+++ internal/backends/mcu/lib.rs
@@ -123,6 +123,7 @@ mod the_backend {
             _items: &mut dyn Iterator<Item = Pin<sixtyfps_corelib::items::ItemRef<'a>>>,
         ) {
         }
+
         fn show_popup(&self, _popup: &ComponentRc, _position: sixtyfps_corelib::graphics::Point) {
             todo!()
         }